build: 関数群を別ファイルに移動
